==> 04_data_analysis/001_snrnaseq_analysis/01_fastqc.qmd <==
---
title: Run FastQC on sample
execute: 
  eval: false
---

We can start by checking some QC metrics of the fasta files.

To count the number of fastq files we can use the following:

```{r}
#| purl: false
ls -lR /gluster/dri02/rdscw/shared/webber/Endo_10X/FASTQ/Set_2/ | grep --count \.fastq.gz$ 
ls -lR /path/to/fastq/set/ | grep --count \.fastq.gz$ 
```

There's 144 in set 1, 244 in set 2, and 656 in set 3.

It's a little graceless, but I'll purl this script for set 1 and then manually create copies and change the array number and directory.

```{r}
#!/bin/bash

#SBATCH -p c_highmem_dri1
#SBATCH --job-name=fastqc_bbb
#SBATCH --ntasks=1
#SBATCH --ntasks-per-node=1
#SBATCH --array=1-144%144
#SBATCH --mem-per-cpu=5000 # memory limit per core
#SBATCH --time=1-00:00 # maximum job time in D-HH:MM
#SBATCH --account=scw1329
#SBATCH -o sbatch_out_%x_%A_%a_%J.txt
#SBATCH -e sbatch_err_%x_%A_%a_%J.txt

## FASTQ file paths and project root variables defined in the following
source /scratch/c.mpmgb/blood-brain-barrier-in-ad/04_data_analysis/001_snrnaseq_analysis/00_hpc_variables.sh

echo "*****************************************************************"
echo "All jobs in this array have:"
echo "- SLURM_ARRAY_JOB_ID: ${SLURM_ARRAY_JOB_ID}"
echo "- SLURM_ARRAY_TASK_COUNT: ${SLURM_ARRAY_TASK_COUNT}"
echo "- SLURM_ARRAY_TASK_MIN: ${SLURM_ARRAY_TASK_MIN}"
echo "- SLURM_ARRAY_TASK_MAX: ${SLURM_ARRAY_TASK_MAX}"
echo "This job in the array has:"
echo "- SLURM_JOB_ID: ${SLURM_JOB_ID}"
echo "- SLURM_ARRAY_TASK_ID: ${SLURM_ARRAY_TASK_ID}"
echo "Run on host: "`hostname`
echo "Number of threads (nproc): "`nproc`
echo "Total memory in GB: "`free -g | grep -oP '\d+' | sed -n 1p`
echo "Used memory in GB: "`free -g | grep -oP '\d+' | sed -n 2p`
echo "Free memory in GB: "`free -g | grep -oP '\d+' | sed -n 3p`
echo "Username: "`whoami`
echo "Started at: "`date`
echo -e "*****************************************************************\n"

OUTPUT_DIR=$PROJECT_ROOT"03_data/990_processed_data/001_snrnaseq/01_fastqc/01_set1"
OUTPUT_DIR2=$PROJECT_ROOT"03_data/990_processed_data/001_snrnaseq/01_fastqc/02_set2"
OUTPUT_DIR3=$PROJECT_ROOT"03_data/990_processed_data/001_snrnaseq/01_fastqc/03_set3"

## Load FastQC module
module load FastQC

#-----------------------------------------------------------------------

mkdir -p $OUTPUT_DIR

N=${SLURM_ARRAY_TASK_ID}

FASTQ_FILE=$(find $INPUT_DIR_b1s1 -mindepth 1 -maxdepth 1 -name '*_001.fastq.gz' | sort | tail -n+${N} | head -1)
fastqc -o $OUTPUT_DIR1 -f fastq --noextract --quiet -t 1 $FASTQ_FILE
 
# FASTQ_FILE=$(find $INPUT_DIR_b1s2 -mindepth 1 -maxdepth 3 -name '*_001.fastq.gz' | sort | tail -n+${N} | head -1)
# fastqc -o $OUTPUT_DIR2 -f fastq --noextract --quiet -t 1 $FASTQ_FILE

# FASTQ_FILE=$(find $INPUT_DIR -mindepth 1 -maxdepth 1 -name '*_001.fastq.gz' | sort | tail -n+${N} | head -1)
# fastqc -o $OUTPUT_DIR3 -f fastq --noextract --quiet -t 1 $FASTQ_FILE


echo -e "\n*****************************************************************"
echo "Finished at: "`date`
echo "*****************************************************************"
```

```{r}
#| purl: false
sbatch --account=<your.project.code.here> -p <your.compute.node> --mail-user <your.email@here> --mail-type END,FAIL 01_fastqc_set1.sh
sbatch --account=<your.project.code.here> -p <your.compute.node> --mail-user <your.email@here> --mail-type END,FAIL 01_fastqc_set2.sh
sbatch --account=<your.project.code.here> -p <your.compute.node> --mail-user <your.email@here> --mail-type END,FAIL 01_fastqc_set3.sh
```
