==> 04_data_analysis/001_snrnaseq_analysis/07_cellranger_aggr.qmd <==
---
title: Aggregate cellranger count reports
execute: 
  eval: false
---

`cellranger count` produces a report per sample, but it's more useful to aggregate these results into one, so we have `cellranger aggr` for this.

This requires a 2 column CSV file (headers of `sample_id,molecule_h5`) with samples ID and the path to the `molecule_info.h5` file.

From the MultiQC reports, it seems the lastest batch if of better quality than the first two, so I'll try using `aggr` on each set separately, and all data combined.

You can find these combined report for all samples [here](../../03_data/990_processed_data/001_snrnaseq/05_cellranger_aggr/cellranger_aggr_all_sets/outs/web_summary.html), set 1 [here](../../03_data/990_processed_data/001_snrnaseq/05_cellranger_aggr/cellranger_aggr_set1/outs/web_summary.html), set 2 [here](../../03_data/990_processed_data/001_snrnaseq/05_cellranger_aggr/cellranger_aggr_set2/outs/web_summary.html), and set 3 [here](../../03_data/990_processed_data/001_snrnaseq/05_cellranger_aggr/cellranger_aggr_set3/outs/web_summary.html).

```{r}
#| purl: false

## Project root variables defined in the following
source /scratch/c.mpmgb/blood-brain-barrier-in-ad/04_data_analysis/001_snrnaseq_analysis/00_hpc_variables.sh

# list the cellranger count output directories
find $PROJECT_ROOT"03_data/990_processed_data/001_snrnaseq/04_cellranger_count/" -maxdepth 2 -mindepth 2 -type d > cellranger_ag.txt 

# use the last directory in the path as the sample id
awk -F'/' '{print $NF "," $0}' cellranger_ag.txt > cellranger_ag2.txt

# append "outs/filtered_gene_bc_matrices" to file paths
sed "s|$|/outs/molecule_info.h5|" cellranger_ag2.txt > cellranger_ag3.txt

# remove the undetermined sample and P_39
grep -v "Undetermined" cellranger_ag3.txt > cellranger_ag4.txt
grep -v "P_39" cellranger_ag4.txt > cellranger_ag5.txt

# separate out the sets
grep "01_set1" cellranger_ag5.txt > cellranger_ag_set1.txt
grep "02_set2" cellranger_ag5.txt > cellranger_ag_set2.txt
grep "03_set3" cellranger_ag5.txt > cellranger_ag_set3.txt

# add the table headers
awk -v line="sample_id,molecule_h5" 'BEGIN{print line}{print}' cellranger_ag_set1.txt > cellranger_aggr_set1.csv
awk -v line="sample_id,molecule_h5" 'BEGIN{print line}{print}' cellranger_ag_set2.txt > cellranger_aggr_set2.csv
awk -v line="sample_id,molecule_h5" 'BEGIN{print line}{print}' cellranger_ag_set3.txt > cellranger_aggr_set3.csv
awk -v line="sample_id,molecule_h5" 'BEGIN{print line}{print}' cellranger_ag5.txt > cellranger_aggr_all_sets.csv

# make a file that has the file paths to use in the array job
echo $PROJECT_ROOT"03_data/990_processed_data/001_snrnaseq/05_cellranger_aggr/cellranger_aggr_set1.csv" > cellranger_aggr_csvs.txt
echo $PROJECT_ROOT"03_data/990_processed_data/001_snrnaseq/05_cellranger_aggr/cellranger_aggr_set2.csv" >> cellranger_aggr_csvs.txt
echo $PROJECT_ROOT"03_data/990_processed_data/001_snrnaseq/05_cellranger_aggr/cellranger_aggr_set3.csv" >> cellranger_aggr_csvs.txt
echo $PROJECT_ROOT"03_data/990_processed_data/001_snrnaseq/05_cellranger_aggr/cellranger_aggr_all_sets.csv" >> cellranger_aggr_csvs.txt

# clean up
rm cellranger_ag.txt cellranger_ag2.txt cellranger_ag3.txt cellranger_ag_set1.txt cellranger_ag_set2.txt cellranger_ag_set3.txt cellranger_ag4.txt cellranger_ag5.txt
```

```{r}
#!/bin/bash

#SBATCH -p c_vhighmem_dri1 ## dev, compute, htc, highmem
#SBATCH --job-name=cellranger_aggr
#SBATCH --ntasks=40
#SBATCH --ntasks-per-node=40
#SBATCH --array=1-4%2
#SBATCH --mem=640G # memory limit per compute node for the job
#SBATCH --time=2-10:00 # maximum job time in D-HH:MM
#SBATCH --account=scw1329
#SBATCH -o /scratch/c.mpmgb/hawk_output/%x_out_%A_%a_%J.txt
#SBATCH -e /scratch/c.mpmgb/hawk_output/%x_err_%A_%a_%J.txt

## Project root variables defined in the following
source /scratch/c.mpmgb/blood-brain-barrier-in-ad/04_data_analysis/001_snrnaseq_analysis/00_hpc_variables.sh

echo "*****************************************************************"
echo "All jobs in this array have:"
echo "- SLURM_ARRAY_JOB_ID: ${SLURM_ARRAY_JOB_ID}"
echo "- SLURM_ARRAY_TASK_COUNT: ${SLURM_ARRAY_TASK_COUNT}"
echo "- SLURM_ARRAY_TASK_MIN: ${SLURM_ARRAY_TASK_MIN}"
echo "- SLURM_ARRAY_TASK_MAX: ${SLURM_ARRAY_TASK_MAX}"
echo "This job in the array has:"
echo "- SLURM_JOB_ID: ${SLURM_JOB_ID}"
echo "- SLURM_ARRAY_TASK_ID: ${SLURM_ARRAY_TASK_ID}"
echo "Run on host: "`hostname`
echo "Number of threads (nproc): "`nproc`
echo "Total memory in GB: "`free -g | grep -oP '\d+' | sed -n 1p`
echo "Used memory in GB: "`free -g | grep -oP '\d+' | sed -n 2p`
echo "Free memory in GB: "`free -g | grep -oP '\d+' | sed -n 3p`
echo "Username: "`whoami`
echo "Started at: "`date`
echo -e "*****************************************************************\n"


## path to input csvs with samples IDs and paths to cellranger count outputs
INPUT_CSVS=$PROJECT_ROOT"03_data/990_processed_data/001_snrnaseq/05_cellranger_aggr/cellranger_aggr_csvs.txt"

## results
OUTPUT_DIR=$PROJECT_ROOT"03_data/990_processed_data/001_snrnaseq/05_cellranger_aggr"

#-----------------------------------------------------------------------
# Cell Ranger aggr

mkdir -p $OUTPUT_DIR

N=${SLURM_ARRAY_TASK_ID}

INPUT_CSV=$(cat $INPUT_CSVS | tail -n+${N} | head -1)
# Subset to last part of the file path
INPUT_CSV_NAME="${INPUT_CSV##*/}"
# Subset string to first "."
SAVE_DIR="${INPUT_CSV_NAME%%.*}"

echo "Input CSV: "$INPUT_CSV
echo "Save dir: "$SAVE_DIR

cd $OUTPUT_DIR

$CELL_RANGER aggr --id=$SAVE_DIR \
--csv=$INPUT_CSV

echo -e "\n*****************************************************************"
echo "Finished at: "`date`
echo "*****************************************************************"
```
