==> 04_data_analysis/001_snrnaseq_analysis/09_scflow_run.qmd <==
---
title: Run SCflow
execute:
  eval: false
bibliography: references.bib
---

# Custom config file

Note that by default the scw config uses the htc queue and we'd rather use our DRI nodes, so I'll overwrite that.

I'm also trying out Nextflow Tower for the first time, it should be nice for keeping an eye on runs, seeing resource usage and troubleshooting any issues.
Note that after I purl this config I'll manually update the Nextflow Tower token on hawk.

```{r}
#| purl: true
executor
{
  name = 'slurm'
  queueSize = 10
  queue = 'c_vhighmem_dri1'
}

params
{
  max_memory = 715.
  GB
  max_cpus = 35
  max_time = 112.
  h
}

tower
{
  accessToken = keyring::key_get("hawk_nextflow_tower")
  enabled = true
}

process
{
  clusterOptions = '--account=scw1329'
  beforeScript = 'module load singularity'
  withName:SCFLOW_MERGE
  {
    memory = 300.
    GB
  }
  withName:SCFLOW_INTEGRATE
  {
    memory = 710.
    GB
    time = 20.
    h
    queue = 'c_vhighmem_dri1'
  }
  withName:SCFLOW_REDUCEDIMS
  {
    memory = 710.
    GB
    queue = 'c_vhighmem_dri1'
  }
  withName:SCFLOW_CLUSTER
  {
    memory = 710.
    GB
    queue = 'c_vhighmem_dri1'
  }
  withName:SCFLOW_REPORTINTEGRATED
  {
    memory = 710.
    GB
    time = 20.
    h
    queue = 'c_vhighmem_dri1'
  }
  withName:SCFLOW_FINALIZE
  {
    memory = 710.
    GB
    queue = 'c_vhighmem_dri1'
  }
  withName:SCFLOW_DIRICHLET
  {
    memory = 710.
    GB
    queue = 'c_vhighmem_dri1'
  }
  withName:SCFLOW_DGE
  {
    memory = 710.
    GB
    queue = 'c_vhighmem_dri1'
  }
  withName:SCFLOW_PLOTREDDIMGENES
  {
    memory = 710.
    GB
    cpus = 10
    time = 20.
    h
    queue = 'c_vhighmem_dri1'
  }
}
```

There is a module one could load on hawk, but it isn't the most recent version of nextflow, so I'm using mamba as per the scw docs

```{r}
mamba
activate
nf - core
```

Note that I'm using the `dev-nf` branch to run, which can be found [here](https://github.com/combiz/nf-core-scflow/tree/dev-nf).

There's two `nextflow run` calls, the first is with both our data and the Yang data, and the latter with just our data in case something goes wrong with that.
[@yang_human_2022]

```{r}
#| eval: false
#| purl: false
nextflow
run
combiz - nf - core - scflow - dev - nf / workflow / -profile
scw - c
03
_data / 990
_processed_data / 001
_snrnaseq / 06
_scflow / pipeline.config - -manifest
03
_data / 990
_processed_data / 001
_snrnaseq / 06
_scflow / scflow_manifest_final.tsv - -input
03
_data / 990
_processed_data / 001
_snrnaseq / 90
_sample_info / sample_metadata.tsv - -reddim_genes_yml
03
_data / 990
_processed_data / 001
_snrnaseq / 06
_scflow / reddim_genes.yml - -ensembl_mappings
03
_data / 990
_processed_data / 001
_snrnaseq / 06
_scflow / ensembl_mappings.tsv - resume - bg > 03
_data / 990
_processed_data / 001
_snrnaseq / 06
_scflow / scflow.log
# version with our samples alone
nextflow
run
combiz - nf - core - scflow - dev - nf / workflow / -profile
scw - c
03
_data / 990
_processed_data / 001
_snrnaseq / 06
_scflow / pipeline.config - -manifest
03
_data / 990
_processed_data / 001
_snrnaseq / 06
_scflow / scflow_manifest.tsv - -input
03
_data / 990
_processed_data / 001
_snrnaseq / 90
_sample_info / sample_metadata_ours.tsv - -reddim_genes_yml
03
_data / 990
_processed_data / 001
_snrnaseq / 06
_scflow / reddim_genes_williams.yml - -ensembl_mappings
03
_data / 990
_processed_data / 001
_snrnaseq / 06
_scflow / ensembl_mappings.tsv - resume - bg > 03
_data / 990
_processed_data / 001
_snrnaseq / 06
_scflow / scflow.log
```
